+++ Sources/TuistKit/GraphMappers/ProjectMapperProvider.swift
@@ -23,6 +23,11 @@ class ProjectMapperProvider: ProjectMapperProviding {
         // Delete current derived
         mappers.append(DeleteDerivedDirectoryProjectMapper())
 
+        // Namespace generator
+        if !config.generationOptions.contains(.disableSynthesizedResourceAccessors) {
+            mappers.append(SynthesizedResourceInterfaceProjectMapper())
+        }
+
         // Support for resources in libraries
         mappers.append(ResourcesProjectMapper())