+++ fixtures/ios_app_with_templates/Project.swift
@@ -10,7 +10,7 @@ let project = Project(name: "App",
                             infoPlist: "Support/Info.plist",
                             sources: ["Sources/**"],
                             resources: [
-                                /* Path to resouces can be defined here */
+                                /* Path to resources can be defined here */
                                 // "Resources/**"
                             ]
                         ),